Experimental let SW play Audio with a content tab
When background service worker need to play a sound, pick an existing
connection, send sound info to the content tab and let it hold <Audio>
and play.

This can avoid creating new Window or Tab to use its DOM.

Problem:

User must interact on the tab first.

    NotAllowedError: play() failed because the user didn't interact with
    the document first. https://goo.gl/xX8pDD

it is reasonable, also the speaker icon seems appear on random tab, is
bad for user.